use destination when pulling remote snapshot list
We run multiple pools and multiple datasets with 1000s of snapshots.  Without the destination specified, the script takes forever to run trying to return the entire list of snapshots when it only needs to know about the snapshots for a specific destination.  This decreases our run time by a factor of 10x at least.